+++ duka-client/registration/translations/index.js
@@ -1,0 +1,70 @@
+module.exports = {
+    'account_number': {
+        'en-ke': 'Please reply with the account number of the farmer\n"0" for new client.',
+        'sw': 'Tafadhali jibu na nambari ya akaunti ya mkulima\n"0" kwa mteja mpya.'
+    },
+    'non_duka_account': {
+        'en-ke': 'You\'ve entered an account number for a non-duka account. Enter your duka account number or 0 to register as a duka client.',
+        'sw': 'Umeingiza nambari ya akaunti isiyo ya duka. Ingiza nambari yako ya akaunti ya duka au \'0\' ili ujiandikishe kama mteja wa duka.'
+    },
+    'already_registered_enter': {
+        'en-ke': 'You\'re already registered for a duka district. Please enter the client\'s invoice ID.',
+        'sw': 'Umesajiliwa kwa wilaya ya duka. Tafadhali ingiza nambari ya Invoice ya mteja'
+    },
+    'outstanding_balance': {
+        'en-ke': 'You have a Duka credit account with an outstanding credit balance of $balance. Please complete your loan in order to take another one.',
+        'sw': 'Una akaunti ya mkopo ya Duka na salio bora la mkopo la X. Tafadhali kamilisha mkopo wako ili uchukue nyingine.'
+    },
+    'confirm_or_try': {
+        'en-ke': '1) To confirm\n2) To try again.',
+        'sw': '1) Kudhibitisha\n2) Kujaribu tena.'
+    },
+    'enter_phone': {
+        'en-ke': 'Enter client Phone number',
+        'sw': 'Ingiza nambari ya simu ya mteja'
+    },
+    'enter_first_name': {
+        'en-ke': 'Please reply with the first name of the client.',
+        'sw': 'Tafadhali jibu na jina la kwanza la mteja.'
+    },
+    'enter_second_name': {
+        'en-ke': 'Please reply with the second name of the client.',
+        'sw': 'Tafadhali jibu na jina la pili la mteja.'
+    },
+    'enter_invoice_id': {
+        'en-ke': 'Please reply with the client\'s Erply Invoice ID',
+        'sw': 'Tafadhali jibu na Kitambulisho cha mteja cha Erply Invoice'
+    },
+    'thanks_for_registering': {
+        'en-ke': 'Thank you for registering with OAF. Your Account Number is $account_number. The Duka team will help you complete your loan.',
+        'sw': 'Asante kwa kusajili na OAF. Nambari yako ya Akaunti ni $account_number. Timu ya Duka itakusaidia kumaliza mkopo wako'
+    },
+    'confirm_or_cancel': {
+        'en-ke': '1) Confirm\n2) Cancel',
+        'sw': '1) Thibitisha\n2) Futa'
+    },
+    'invoice_id_confirm': {
+        'en-ke': 'You entered invoice ID $invoice_id\n',
+        'sw': 'Uliingiza nambari $invoice_id ya invoice. Ingiza\n'
+    },
+    'national_id_confirm': {
+        'en-ke': 'You entered $national_id ID\n',
+        'sw': 'Ingiza nambari $national_id ya kitamburisho. Ingiza\n'
+    },
+    'enter_national_id': {
+        'en-ke': 'Enter client national ID\n',
+        'sw': 'Ingiza nambari ya kitambulisho cha kitaifa cha mteja\n'
+    },
+    'phone_number_confirm': {
+        'en-ke': 'You entered phone number $phone_number\n',
+        'sw': 'Uliingiza nambari $phone_number ya simu. Ingiza\n'
+    },
+    'first_sencond_name_confirm': {
+        'en-ke': 'Enroll client $first_name $second_name\n',
+        'sw': 'Sajili mteja $first_name $second_name\n'
+    },
+    'successfull_transaction': {
+        'en-ke': 'You have successfully recorded your transaction.',
+        'sw': 'Umekamilisha mafanikio ya ununuzi wako.'
+    }
+};