+++ CHANGELOG.md
@@ -9,9 +9,16 @@ All notable changes to this project will be documented in this file.
 * [SER-219](https://oneacrefund.atlassian.net/browse/SER-219) As a client, I want to be able to receive a maize recommendation, so that I know which variety is optimal for me
 
 ## 2020-09-23
-### Added
+## Added
 
 * [SER-209](https://oneacrefund.atlassian.net/browse/SER-209) As a client, I want to add avocado trees to my order, so that I can plant a high-impact product
+* [SER-206](https://oneacrefund.atlassian.net/browse/SER-206) As a credit officer, I want to register clients so that their purchase can be recorded in Roster
+* [SER-221](https://oneacrefund.atlassian.net/browse/SER-221) Create a data table in Telerivet that stores a client's account number and invoice ID
+
+## Changed
+
+* [SER-9](https://oneacrefund.atlassian.net/browse/SER-9)As a GL, I want to be able to top up a client's order via USSD, so that I can adjust orders prior to distribution
+Before one sms was being sent to the phone being used for the session. This changes it to send another sms on the client's active phone stored in Roster. It also change the content of the sms to include the ordered products 
 
 
 ## 2020-09-21